Update patch set 3
Patch Set 3:

including gradle-wapper vs make Gradle a requirement - it's something I've been thinking on and I think the wrapper is actually nice both for CI, as it will be easy to add jenkins instances if needed without installing Gradle on the machine, and for developers that won't have to install anythings besides the JVM in order to start working.

Assuming I'll be able to add the wrapper in one place for all the sub-projects (when we'll have multi-projects structure), will you still won't like to use it?
BTW in Gradle tutorial they are suggesting to add the wrapper to the version control:
http://www.gradle.org/docs/current/userguide/gradle_wrapper.html

Regarding the bug in OsVersion - good point. I'll fix it & other bug on 10.10 that was opened yesterday:
https://github.com/mucommander/mucommander/issues/3
and install mac os x 10.10 to verify both of them

Patch-set: 3
Label: Code-Review=0